\name{addFeatureData-methods}
\docType{methods}
\alias{addFeatureData}
\alias{addFeatureData-methods}
\alias{addFeatureData,eSet-method}
\title{Method for Function addFeatureData}
\description{
Methods to add SNP-level annotation to the featureData slot in eSets.
Feature-level annotation can include chromosome, physical position,
allele, strand (sense/anti-sense), etc.
}
\section{Methods}{
\describe{
\item{object = "eSet"}{Any object that is an instance of a class extending eSet}
}}
\keyword{methods}
