\name{alleleA-methods}
\docType{methods}
\alias{alleleA}
\alias{alleleA-methods}
\alias{alleleA,AnnotatedDataFrame-method}
\alias{alleleA,SnpCallSet-method}
\alias{alleleA,SnpCopyNumberSet-method}
\alias{alleleA,oligoSnpSet-method}
\alias{alleleA,AnnotatedSnpSet-method}
\title{ Accessor for allele A of a SNP}
\description{
 Accessors \code{alleleA}  retrieve allele A of a SNP
}
\section{Methods}{
\describe{
  \item{object = "AnnotatedDataFrame"}{retrieves allele A of a SNP}
  \item{object = "SnpCallSet"}{retrieves allele A of a SNP}
  \item{object = "SnpCopyNumberSet"}{retrieves allele A of a SNP}
  \item{object = "oligoSnpSet"}{retrieves allele A of a SNP}
  \item{object = "AnnotatedSnpSet"}{retrieves allele A of a SNP}
}}
\keyword{methods}

