\name{alleleA}
\alias{alleleA}
\alias{alleleA,AnnotatedDataFrame-method}
\alias{alleleA,SnpLevelSet-method}
\title{Accessor for the A allele}
\description{
}
\usage{
alleleA(object)
}
\arguments{
  \item{object}{one of the classes defined in \code{SNPchip}}
}
\details{
}
\value{
  Returns a vector of the A alleles 
}
\author{Robert Scharpf}
\note{
}
\seealso{\code{\link{alleleA}}}
\examples{
}
\keyword{methods}

