\name{alleleB-methods}
\docType{methods}
\alias{alleleB-methods}
\alias{alleleB,AnnotatedDataFrame-method}
\alias{alleleB,SnpCallSet-method}
\alias{alleleB,SnpCopyNumberSet-method}
\alias{alleleB,oligoSnpSet-method}
\title{Accessor for allele B of a SNP}
\description{
  Accessors \code{alleleB} retrieve allele B of a SNP
}
\section{Methods}{
\describe{
  \item{object = "AnnotatedDataFrame"}{retrieves allele B of a SNP}
  \item{object = "SnpCallSet"}{retrieves allele B of a SNP }
  \item{object = "SnpCopyNumberSet"}{retrieves allele B of a SNP }
  \item{object = "oligoSnpSet"}{retrieves allele B of a SNP }
  \item{object = "AnnotatedSnpSet"}{retrieves allele B of a SNP }
}}
\keyword{methods}

