\name{alleleB}
\alias{alleleB}
\alias{alleleB,eSet-method}
\alias{alleleB,AnnotatedDataFrame-method}
\title{Accessor for the B allele}
\description{
}
\usage{
alleleB(object)
}
\arguments{
  \item{object}{one of the classes defined in \code{SNPchip}}
}
\details{
}
\value{
  Returns a vector of the B alleles 
}
\author{Robert Scharpf}
\note{
}
\seealso{\code{\link{alleleB}}}
\examples{
}
\keyword{methods}
