\name{allPlots-methods}
\docType{methods}
\alias{allPlots}
\title{Extracts graphical parameters to be used by all plotting devices}
\description{
}
\section{Methods}{
\describe{
\item{object = "ParESet"}{class containing graphical parameters for
  high-throughput SNP datasets}
}}
\keyword{methods}

