\name{AnnotatedSnpCallSetList-class}
\docType{class}
\alias{AnnotatedSnpCallSetList-class}
\alias{coerce,AnnotatedSnpCallSetList,AnnotatedSnpCallSet-method}
\alias{snpSetList,AnnotatedSnpCallSetList-method}

\title{A list with elements AnnotatedSnpCallSet.}
\description{A list with elements AnnotatedSnpCallSet. Each element is
  typically an AnnotatedSnpCallSet of one chromosome.}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("AnnotatedSnpCallSetList", ...)}.
}
\section{Slots}{
  \describe{
    \item{\code{snpSetList}:}{Object of class \code{"list"}}
  }
}
\section{Methods}{
  No methods defined with class "AnnotatedSnpCallSetList" in the signature.
}
\author{Robert Scharpf}
\note{}
\seealso{
  \code{\linkS4class{AnnotatedSnpCallSet}}
}
\examples{
}
\keyword{classes}
