\name{annSnpset}
\alias{annSnpset}
\docType{data}
\title{AnnotatedSnpSet object}
\description{
AnnotatedSnpSet object from Hapmap data.
}
\usage{data(annSnpset)}
\format{
  The format is:
 chr "annSnpset"
}
\details{

}
\source{
}
\references{
}
\examples{
data(annSnpset)
annSnpset
}
\keyword{datasets}
