\name{calculateCopyNumber-methods}
\docType{methods}
\alias{calculateCopyNumber}
\alias{calculateCopyNumber-methods}
\alias{calculateCopyNumber,SnpQSet-method}
\title{Extracts copy number estimates from thetas in SnpQSet object}

\description{
  Extracts copy number estimates from thetas in \code{SnpQSet} objects.
}

\section{Methods}{
\describe{


}}
\keyword{methods}
\keyword{manip}

