\name{chromosomeAnnotation-methods}
\docType{methods}
\alias{chromosomeAnnotation-methods}
\alias{chromosomeAnnotation,AnnotatedSnpSet-method}
\alias{chromosomeAnnotation,AnnotatedSnpCopyNumberSet-method}
\alias{chromosomeAnnotation,AnnotatedSnpCallSet-method}
\alias{chromosomeAnnotation<--methods}
\alias{chromosomeAnnotation<-,AnnotatedSnpCallSet,data.frame-method}
\alias{chromosomeAnnotation<-,AnnotatedSnpCopyNumberSet,data.frame-method}
\alias{chromosomeAnnotation<-,AnnotatedSnpSet,data.frame-method}
\title{chromosome-level annotation for SnpSet objects}
\description{
chromosome-level annotation for SnpSet objects
}
\section{Methods}{
\describe{

\item{object = "AnnotatedSnpCallSet", value = "data.frame"}{replacement
  method for chromosomeAnnotation slot}

\item{object = "AnnotatedSnpCopyNumberSet", value = "data.frame"}{replacement
  method for chromosomeAnnotation slot}

\item{object = "AnnotatedSnpSet", value = "data.frame"}{replacement
  method for chromosomeAnnotation slot}

\item{object = "AnnotatedSnpCopyNumberSet"}{returns all chromosome-level
annotation in object of class AnnotatedSnpCopyNumberSet}

\item{object = "AnnotatedSnpCallSet"}{returns all chromosome-level
annotation in object of class AnnotatedSnpCopyNumberSet}

\item{object = "AnnotatedSnpSet"}{  returns all chromosome-level
annotation in object of class AnnotatedSnpCopyNumberSet}
}}
\keyword{methods}

