\name{chromosomeAnnotation}
\alias{chromosomeAnnotation}
\docType{data}
\title{chromosome annotation}
\description{
Contains  information on chromosomes 1-22, X and Y.
}
\usage{data(chromosomeAnnotation)}
\format{
  A data frame with 24 observations on the following 3 variables.
  \describe{
    \item{\code{centromereStart}}{a numeric vector}
    \item{\code{centromereEnd}}{a numeric vector}
    \item{\code{chromosomeSize}}{a numeric vector}
  }
}
\examples{
data(chromosomeAnnotation)
}
\keyword{datasets}
