\name{classes-class}
\docType{class}
\alias{oligoSnpSet-class}
\alias{SnpCopyNumberSet-class}
\alias{SnpCallSet-class}
\alias{SnpQSet-class}
\title{Imported classes from oligo}
\description{}
\references{}
\author{}
\note{}
\seealso{
  \code{\link[oligo]{oligoSnpSet-class}}, \code{\link{par}}, \code{\link[graphics]{layout}}
}
\keyword{classes}

