\name{copyNumber-methods}
\docType{methods}
\alias{copyNumber-methods}
\alias{copyNumber,SnpCopyNumberSet-method}
\alias{copyNumber,SnpQSet-method}
\title{Methods for Function copyNumber in Package SNPscan}
\description{
Methods for function \code{copyNumber} 
}
\section{Methods}{
\describe{

\item{object = "SnpCopyNumberSet"}{accessor for copy number in object of
class SnpCopyNumberSet}

\item{object = "SnpQSet"}{Calculates copy number by taking the row means
of the 'A' values in an object of class SnpQSet.}
}}
\keyword{methods}

