\name{copyNumber-methods}
\docType{methods}
\alias{copyNumber-methods}
\alias{copyNumber,SnpQSet-method}
\title{Extracts copy number estimates}

\description{
  Extracts copy number estimates from \code{AnnotatedSnpSet} objects.
}

\section{Methods}{
\describe{

  \item{copyNumber}{\code{\signature(object="SnpQSet")}: accessor for matrix of
  copy number estimates.  Rows correspond to SNPs and columns to
  samples.}

}}
\keyword{methods}
\keyword{manip}

