\name{crlmmOut}
\alias{crlmmOut}
\docType{data}
\title{A data object created by the oligo vignette}

\description{
  
  This data object was created by the oligo vignette from 3 Hapmap
  samples and then subset to include only the SNPs on chromosome 1.  See
  the oligo vignette for additional details.
  
}

\usage{data(crlmmOut)}
\examples{
data(crlmmOut)
}
\keyword{datasets}
