\name{dbSnpId-methods}
\docType{methods}
\alias{dbSnpId-methods}
\alias{dbSnpId,SnpCallSet-method}
\alias{dbSnpId,AnnotatedDataFrame-method}
\alias{dbSnpId,SnpCopyNumberSet-method}
\alias{dbSnpId,oligoSnpSet-method}
\alias{dbSnpId,AnnotatedSnpSet-method}
\title{Accessor for dbSNP identifier}
\description{
Accessor for dbSNP identifier
}
\section{Methods}{
\describe{

  \item{object = "SnpCallSet"}{returns dbSNP identifier for each SNP in
    object of class SnpCallSet  }

  \item{object = "AnnotatedDataFrame"}{returns dbSNP identifier for each SNP}

  \item{object = "SnpCopyNumberSet"}{returns dbSNP identifier for each SNP in
    object of class SnpCopyNumberSet}

  \item{object = "oligoSnpSet"}{returns dbSNP identifier for each SNP in
    object of class oligoSnpSet}

  \item{object = "AnnotatedSnpSet"}{returns dbSNP identifier for each SNP in
    object of class AnnotatedSnpSet}
}}
\keyword{methods}

