\name{enzyme-methods}
\docType{methods}
\alias{enzyme-methods}
\alias{enzyme,AnnotatedDataFrame-method}
\alias{enzyme,SnpCopyNumberSet-method}
\alias{enzyme,SnpCallSet-method}
\alias{enzyme,oligoSnpSet-method}
\title{Enzyme for DNA digestion}
\description{
  An indicator for which chip the SNP originated from.
}
\section{Methods}{
  \describe{
    
    \item{object = "AnnotatedDataFrame"}{enzyme used to digest the DNA}
  
    \item{object = "SnpCopyNumberSet"}{ enzyme used to digest the DNA }
    
    \item{object = "SnpCallSet"}{ enzyme used to digest the DNA }

    \item{object = "oligoSnpSet"}{ enzyme used to digest the DNA }

    \item{object = "AnnotatedSnpSet"}{ enzyme used to digest the DNA }
}}
\keyword{methods}

