\name{fragmentLength}
\alias{fragmentLength}
\alias{fragmentLength,AnnotatedDataFrame-method}
\alias{fragmentLength,SnpLevelSet-method}
\title{Accessor for the fragment length}
\description{
  Accessor for the length of the PCR fragment on which the SNP resides.
}
\usage{
fragmentLength(object)
}
\arguments{
  \item{object}{Object, possibly derived from class \code{SnpCallSet} or
    \code{SnpCopyNumberSet}}
}


\value{ Length of the PCR fragment on which the SNP resides after
  enzymatic digestion.  }

\author{R. Scharpf}
\keyword{manip}

