\name{getPar-methods}
\docType{methods}
\alias{getPar}
\alias{getPar-methods}
\alias{getPar,SnpLevelSet-method}
\title{Computes default graphical parameters }
\description{
Computes default graphical parameters for high throughput SNP
  objects inheriting from \code{eSet}
}
\section{Methods}{
\describe{
\item{object = "eSet"}{ object inheriting from class \code{eSet}}
}}
\keyword{methods}

