\name{hapmap}
\alias{hapmap}
\docType{data}
\title{Data from the hapmap project}
\description{
Data obtained from the hapmap project and pre-processed by CRLMM to
produce estimates of copy number and genotype calls.  Due to size
limitations, we only included every 10th SNP from the Affy's 50k Xba chip.
}
\usage{data(hapmap)}
\author{R. Scharpf}
\examples{
data(hapmap)
}
\keyword{datasets}
