\name{initialize-methods}
\docType{methods}
\alias{initialize-methods}
\alias{initialize,SnpCallSet-method}
\alias{initialize,AnnotatedSnpCallSet-method}
\alias{initialize,SnpCopyNumberSet-method}
\alias{initialize,AnnotatedSnpCopyNumberSet-method}
\alias{initialize,oligoSnpSet-method}
\alias{initialize,AnnotatedSnpSet-method}
\alias{initialize,AnnotatedSnpSetList-method}
\title{Generates instances of SnpSet objects}
\description{
Generates instances of SnpSet objects
}
\section{Methods}{
\describe{

\item{.Object = "SnpCallSet"}{ generates an instance of an object of class SnpCallSet}

\item{.Object = "AnnotatedSnpCallSet"}{ generates an instance of an object of class AnnotatedSnpCallSet}

\item{.Object = "SnpCopyNumberSet"}{ generates an instance of an object
  of class SnpCopyNumberSet }

\item{.Object = "AnnotatedSnpCopyNumberSet"}{ generates an instance of an object of class AnnotatedSnpCopyNumberSet}

\item{.Object = "oligoSnpSet"}{ generates an instance of an object of
  class oligoSnpSet }

\item{.Object = "AnnotatedSnpSet"}{ generates an instance of an object of class AnnotatedSnpSet}

\item{.Object = "AnnotatedSnpSetList"}{ generates an instance of an object of class AnnotatedSnpSetList}
}}
\keyword{methods}

