\name{integer2chromosome}
\alias{integer2chromosome}
\title{Converts integer representation of chromosome number back to letter representation}
\description{
	Coerces integer representation for chromosome in the pd.annotation packages to character strings
}
\usage{
	integer2chromosome(chrom)
}
\arguments{
	\item{chrom}{chromosome}
}
\value{
	  \code{character}
}
\examples{
	integer2chromosome(1:26)
}	
\author{R. Scharpf}
\keyword{manip}
