\name{plotChromosome-methods}
\docType{methods}
\alias{plotChromosome-methods}
\alias{plotChromosome,AnnotatedSnpSet-method}
\title{Plots copy number and genotype calls versus physical position for
1 chromosome}
\description{
Plots copy number and genotype calls versus physical position for
1 chromosome
}
\section{Methods}{
\describe{

\item{object = "AnnotatedSnpSet"}{Object of the class AnnotatedSnpSet
  must contain only 1 chromosome}
}}
\keyword{methods}

