\name{plotChromosome}
\alias{plotChromosome}
\title{Plots copy number color-coded by genotype (optionally) for a
  single chromosome}
\description{
}
\usage{
plotChromosome(object, ...)
}
\arguments{
  \item{object}{\code{AnnotatedSnpSet} or \code{AnnotatedSnpCopyNumberSet}}
  \item{\dots}{Additional arguments}
}
\details{
}
\value{
}
\author{Robert Scharpf}
\note{
}
\seealso{\code{\link{plotSnp}}}
\examples{
}
\keyword{methods}

