\name{plotCytoband-methods}
\docType{methods}
\alias{plotCytoband}
\alias{plotCytoband-methods}
\alias{plotCytoband,AnnotatedSnpSet-method}
\title{Plots cytoband}
\description{
Plots the cytoband for a single chromosome.
}
\section{Methods}{
\describe{

\item{object = "AnnotatedSnpSet"}{ an object of class AnnotatedSnpSet
  that contains only 1 chromosome.}
}}
\keyword{methods}
