\name{plotPredictions}
\alias{plotPredictions}
\title{Plots the predictions from the HMM}
\description{
Plots the predictions from the HMM.  This function is usually not called directly.
}
\usage{
plotPredictions(object, op)
}
\arguments{
  \item{object}{object of class \code{HmmPredict}}
  \item{op}{an object inheriting from \code{ParESet}}
}
\value{
  Nothing
}
\author{R. Scharpf}
\examples{
##See the VanillaICE vignette
}
\keyword{aplot}

