\name{position-methods}
\docType{methods}
\alias{position-methods}
\alias{position,SnpCallSet-method}
\alias{position,AnnotatedDataFrame-method}
\alias{position,SnpCopyNumberSet-method}
\alias{position,oligoSnpSet-method}
\alias{position,AnnotatedSnpSet-method}
\title{Methods for Function position}
\description{
 Methods for function \code{position} 
}
\section{Methods}{
  \describe{

    \item{object = "SnpCallSet"}{Returns physical position of SNP along the
      chromosome in numbers of basepairs from the 5' end (an integer)}

    \item{object = "AnnotatedDataFrame"}{Returns physical position of SNP along the
      chromosome in numbers of basepairs from the 5' end (an integer)}

    \item{object = "SnpCopyNumberSet"}{Returns physical position of SNP along the
      chromosome in numbers of basepairs from the 5' end (an integer)}

    \item{object = "oligoSnpSet"}{Returns physical position of SNP along the
      chromosome in numbers of basepairs from the 5' end (an integer)}

    \item{object = "AnnotatedSnpSet"}{Returns physical position of SNP along the
      chromosome in numbers of basepairs from the 5' end (an integer)}
}}
\keyword{methods}
