\name{probeSetId-methods}
\docType{methods}
\alias{probeSetId-methods}
\alias{probeSetId,SnpCallSet-method}
\alias{probeSetId,AnnotatedDataFrame-method}
\alias{probeSetId,SnpCopyNumberSet-method}
\alias{probeSetId,oligoSnpSet-method}
\alias{probeSetId,AnnotatedSnpSet-method}
\title{ Accessor for Affymetrix probe-set identifiers }
\description{
 Accessor for Affymetrix probe-set identifiers 
}
\section{Methods}{
  \describe{

    \item{object = "SnpCallSet"}{accessor for Affymetrix probe-set identifiers}

    \item{object = "AnnotatedDataFrame"}{accessor for Affymetrix probe-set identifiers}

    \item{object = "SnpCopyNumberSet"}{accessor for Affymetrix probe-set identifiers  }

    \item{object = "oligoSnpSet"}{accessor for Affymetrix probe-set identifiers  }

    \item{object = "AnnotatedSnpSet"}{accessor for Affymetrix probe-set identifiers  }
}}
\keyword{methods}

