\name{sample.snpset}
\alias{sample.snpset}
\docType{data}
\title{Example object of class oligoSnpSet}
\description{
  Example object of class \code{oligoSnpSet} created from 5 HapMap
  samples assayed on the Affymetrix 100k platform.
}
\usage{data(sample.snpset)}
\details{
}
\source{
}
\references{
}
\examples{
data(sample.snpset)
sample.snpset
}
\keyword{datasets}
