\name{selectSomeIndex}
\alias{selectSomeIndex}
\alias{selectSomeIndex,data.frame-method}
\title{Method from Biobase for selecting indices }
\description{
  Method from Biobase for selecting indices in a data.frame.  This
  method is not currently exported in Biobase. 
}
\usage{
selectSomeIndex(object, ...)
}
\arguments{
  \item{object}{a \code{data.frame}}
  \item{\dots}{additional arguments}
}
\details{

}
\value{
   a list of indices
}
\references{}
\author{}
\note{ 
}
\seealso{}
\examples{
}
\keyword{methods}

