\name{show-methods}
\docType{methods}
\alias{show-methods}
\alias{show,AnnotatedSnpSet-method}
\title{show method for SNPscan objects}
\description{
show method for SNPscan objects
}
\section{Methods}{
\describe{
\item{object = "AnnotatedSnpSet"}{ show method for AnnotatedSnpSet}
}}
\keyword{methods}
