\name{showSummary}
\alias{showSummary}
\title{Displays descriptive statistics for a sample.}
\description{
  Displays descriptive statistics for a sample as a legend.
}
\usage{
showSummary(object, where, bty, legend.panel, cex, col, digits)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{Object of class \code{oligoSnpSet}}
  \item{where}{Character string indicating where to put the legend.  See
  \code{legend}}
  \item{bty}{See \code{bty}}
  \item{legend.panel}{Logical for plotting the legend on a separate panel.}
  \item{cex}{See \code{par}}
  \item{col}{See \code{par}}
  \item{digits}{Argument to \code{round}}
}
\details{
}
\value{
}
\references{}
\author{Robert Scharpf}
\note{
}
\seealso{\code{\link[graphics]{legend}}, \code{\link[base:Round]{round}},
  \code{\link[graphics]{par}}} 
\examples{
}
\keyword{aplot}

