\name{snpSetList-methods}
\docType{methods}
\alias{snpSetList}
\title{Method for function snpSetList in Package SNPscan}
\description{
Method for function \code{snpSetList} in Package SNPscan
}
\section{Methods}{
\describe{

\item{object = "AnnotatedSnpSetList"}{accessor for retrieving list of AnnotatedSnpSets}
}}
\keyword{methods}
