\name{summary-methods}
\docType{methods}
\alias{summary-methods}
\alias{summary,ANY-method}
\alias{summary,SnpCallSet-method}
\alias{summary,AnnotatedSnpSet-method}
\title{Methods for Function summary}
\description{
Methods for function \code{summary} 
}
\section{Methods}{
\describe{

\item{object = "ANY"}{describe this method here }

\item{object = "SnpCallSet"}{calculates percent homozygous calls, and
  percent heterozygous calls for each chromosome, and then aggregated
  over all autosomes.}

\item{object = "AnnotatedSnpSet"}{calculates average copy number,
  standard deviation of copy number, percent homozygous calls, and
  percent heterozygous calls for each chromosome, and then aggregated
  over all autosomes.}
  }} 
\keyword{methods}

