\name{Accessors}
\alias{probeSetId}
\alias{dbSnpId}
\title{Accessors for AnnotatedSnpSet objects}
\description{Set of accessors for AnnotatedDataFrame objects.}
\usage{
probeSetId(object)
dbSnpId(object)
}
\arguments{
  \item{object}{Object derived from class \code{AnnotatedDataFrame}}
}
\value{
 \code{dbSnpId} returns the dbSNP identifier;
 \code{probeSetId} returns the Affymetrix identifier.
}
\examples{
}
\keyword{manip}% at least one, from doc/KEYWORDS
