\name{alleleA}
\alias{alleleA}
\alias{alleleA,AnnotatedDataFrame-method}
\alias{alleleA,SnpLevelSet-method}
\title{Accessor for the A allele}
\description{
  Accessor for the A allele
}
\usage{
alleleA(object)
}
\arguments{
  \item{object}{one of the classes defined in \code{SNPchip}}
}
\value{
  Returns a vector of the A alleles 
}
\author{R. Scharpf}

\seealso{\code{\link{alleleA}}}

\keyword{methods}

