\name{alleleB}
\alias{alleleB}
\alias{alleleB,SnpLevelSet-method}
\alias{alleleB,AnnotatedDataFrame-method}
\title{Accessor for the B allele}
\description{Accessor for the B allele
}
\usage{
alleleB(object)
}
\arguments{
  \item{object}{one of the classes defined in \code{SNPchip}}
}
\value{
  Returns a vector of the B alleles 
}
\author{R. Scharpf}

\seealso{\code{\link{alleleB}}}

\keyword{methods}
