\name{chromosome-methods}
\docType{methods}
\alias{chromosome}
\alias{chromosome-methods}
\alias{chromosome,oligoSnpSet-method}
\alias{chromosome,SnpCallSet-method}
\alias{chromosome,SnpCopyNumberSet-method}
\alias{chromosome,AnnotatedDataFrame-method}

\title{Accessor for chromosome SNP annotation}
\description{
  
  Extracts character string of chromosome number for each SNP in the
  \code{AnnotatedDataFrame} object.

}


\section{Methods}{
\describe{
\item{chromosome}{\code{signature(object="AnnotatedDataFrame")}:
  Character vector indicating the chromosome on which the SNPs in the
  \code{AnnotatedDataFrame} object are located.}
}}
\keyword{methods}
\keyword{manip}

