\name{chromosomeSize}
\alias{chromosomeSize}
\title{Get size of chromosome in number of base pairs.}
\description{
  Size of chromosome in base pairs.
}
\usage{
chromosomeSize(chromosome)
}

\arguments{
  \item{chromosome}{character string for the chromosome, e.g., "1", ...,
    "22", "X", or "Y"}
}

\value{
  Numerical vector.
}

\author{Robert Scharpf}
\examples{
chromosomeSize("1")
}
\keyword{manip}

