\name{chromosomeSize}
\alias{chromosomeSize}
\title{Get size of chromosome in number of base pairs.}
\description{
  Size of chromosome in base pairs.
}
\usage{
chromosomeSize(chromosome)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{chromosome}{character string for the chromosome, e.g., "1", ...,
    "22", "X", or "Y"}
}
\details{
}
\value{
  Numerical vector.
}
\references{}
\author{Robert Scharpf}
\note{
}
\seealso{}
\examples{
chromosomeSize("1")
}
\keyword{manip}

