\name{coerce-methods}
\docType{methods}
\alias{coerce-methods}
\alias{coerce,oligoSnpSet,AnnotatedSnpSet-method}
\alias{coerce,AnnotatedSnpSet,AnnotatedSnpSetList-method}
\alias{coerce,AnnotatedSnpSetList,AnnotatedSnpSet-method}
\title{Methods for Function coerce }
\description{
Methods for Function coerce
}
\section{Methods}{
\describe{

\item{from = "oligoSnpSet", to = "AnnotatedSnpSet"}{coerces oligoSnpSet
  to AnnotatedSnpSet}

\item{from = "AnnotatedSnpSet", to = "AnnotatedSnpSetList"}{coerces
  object of class AnnotatedSnpSet to AnnotatedSnpSetList}

\item{from = "AnnotatedSnpSetList", to = "AnnotatedSnpSet"}{coerces
  object of class AnnotatedSnpSetList to AnnotatedSnpSet}
}}
\keyword{methods}

