\name{coerce-methods}
\docType{methods}
\alias{coerce-methods}
\alias{coerce,oligoSnpSet,AnnotatedSnpSet-method}
\title{Methods for Function coerce }
\description{
Methods for Function coerce
}
\section{Methods}{
\describe{

\item{from = "oligoSnpSet", to = "AnnotatedSnpSet"}{coerces oligoSnpSet
  to AnnotatedSnpSet}
}}
\keyword{methods}

