\name{coerce-methods}
\docType{methods}
\alias{coerce-methods}
\alias{coerce,oligoSnpSet,SnpCallSet-method}
\alias{coerce,oligoSnpSet,SnpCopyNumberSet-method}
\alias{coerce,AnnotatedSnpSet,oligoSnpSet-method}
\alias{coerce,AnnotatedSnpCallSet,SnpCallSet-method}
\alias{coerce,AnnotatedSnpCopyNumberSet,SnpCopyNumberSet-method}
\title{Methods for Function coerce }
\description{
Methods for Function coerce
}
\section{Methods}{
\describe{
  Coercion methods for classes defined in SNPchip
}}
\keyword{methods}

