\name{initialize-methods}
\docType{methods}
\alias{initialize-methods}
\alias{initialize,oligoSnpSet-method}

\title{Generates instances of SnpSet objects}
\description{
Generates instances of SnpSet objects
}
\section{Methods}{
\describe{

\item{.Object = "oligoSnpSet"}{ generates an instance of an object of
  class oligoSnpSet }

}
}
\keyword{methods}

