\name{position-methods}
\docType{methods}
\alias{position}
\alias{position,AnnotatedDataFrame-method}
\alias{position,SnpLevelSet-method}

\title{Physical position of SNP on chromosome}
\description{
  The physical position of a SNP is the number of nucleotide base pairs
  of a SNP from the 5 prime end of the chromosome.
}
\section{Methods}{
\describe{

  \item{position}{\code{signature(object="AnnotatedDataFrame")}: Returns
  the physical position for each SNP in the \code{AnnotatedDataFrame}
  object.}

}
}
\keyword{methods}
\keyword{manip}

