\name{show-methods}
\docType{methods}
\alias{show-methods}
\alias{show,oligoSnpSet-method}
\alias{show,ParESet-method}
\title{show methods}
\description{
Show methods
}
\section{Methods}{
\describe{
  \item{object = "oligoSnpSet"}{ show method for \code{oligoSnpSet} objects}
  \item{object = "ParESet"}{show method for objects inheriting from the
    \code{ParESet} class}
}}
\keyword{methods}
