\name{show-methods}
\docType{methods}
\alias{show-methods}
\alias{show,oligoSnpSet-method}
\alias{show,ParESet-method}
\alias{show,ParSnpCopyNumberSet-method}
\alias{show,ParSnpCallSet-method}
\alias{show,ParSnpSet-method}
\title{show methods}
\description{
Show methods
}
\section{Methods}{
\describe{
  \item{object = "oligoSnpSet"}{ show method for \code{oligoSnpSet} objects}
  \item{object = "ParESet"}{show method for objects inheriting from the
    \code{ParESet} class}
}}
\details{
  The show methods can be used to plot instances of the ParESet
  classes.  An object of the class is created using \code{plotSnp}.
}
\seealso{\code{\link{plotSnp}}}
\keyword{methods}
