\name{smoothSnp-methods}
\docType{methods}
\alias{smoothSnp-methods}
\alias{smoothSnp,AnnotatedSnpSet-method}
\alias{smoothSnp}
\title{Methods for Function smoothSnp in Package SNPscan }
\description{
Methods for function \code{smoothSnp} in Package SNPscan
}
\section{Methods}{
\describe{

\item{object = "AnnotatedSnpSet"}{object of class AnnotatedSnpSet}
}}
\keyword{methods}

